requirements.txt
Django==2.2.7
pytz==2019.3
sqlparse==0.3.0
Pillow==6.2.1
djangorestframework==3.10.3
